Move api from draft to public
