+++ ydb/public/api/grpc/draft/ya.make
@@ -21,7 +21,6 @@ SRCS(
     ydb_replication_v1.proto
     ydb_sqs_topic_v1.proto
     ydb_tablet_v1.proto
-    ydb_test_shard_v1.proto
     ydb_view_v1.proto
     ydb_ymq_v1.proto
 )
