+++ ydb/public/sdk/cpp/src/client/draft/ya.make
@@ -6,7 +6,6 @@ SRCS(
     ydb_dynamic_config.cpp
     ydb_replication.cpp
     ydb_scripting.cpp
-    ydb_test_shard.cpp
     ydb_view.cpp
 )
