+++ ydb/public/sdk/cpp/src/client/ya.make
@@ -42,6 +42,7 @@ RECURSE(
     table
     table/impl
     table/query_stats
+    test_shard
     topic
     topic/codecs
     topic/impl
