+++ ydb/services/test_shard/grpc_service.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <ydb/public/api/grpc/draft/ydb_test_shard_v1.grpc.pb.h>
+#include <ydb/public/api/grpc/ydb_test_shard_v1.grpc.pb.h>
 
 #include <ydb/library/grpc/server/grpc_server.h>
 #include <ydb/library/actors/core/actorid.h>
