+++ ydb/services/test_shard/grpc_service_ut.cpp
@@ -5,7 +5,7 @@
 #include <ydb/core/tx/scheme_cache/scheme_cache.h>
 
 #include <ydb/public/api/grpc/ydb_scheme_v1.grpc.pb.h>
-#include <ydb/public/api/grpc/draft/ydb_test_shard_v1.grpc.pb.h>
+#include <ydb/public/api/grpc/ydb_test_shard_v1.grpc.pb.h>
 
 #include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/resources/ydb_resources.h>
